Add files via upload
Author:     Koto Andrew Omiloli
Title:      8-Puzzle Problem Implementation Using A* search
Date:       11-28-2024

total cost (fval) = h + exp(-5*g) where
h = number of misplaced tiles
g = number of layer current state belongs to

The program solves the 8-puzzle tile game for any given solvable state. It checks an input start state for the number of inversions to determine if the puzzle is solvable. If the inversion count is even, it returns the sequence of steps to reach a desired goal state.  Please cite this work accordingly.